pyproject.toml
[tool.poetry]
name = "rag_customer_support"
version = "0.1.0"
description = ""
authors = ["Leandro A. Hidalgo <le.hidalgot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
pyyaml = "^6.0.2"
transformers = "^4.45.2"
langchain-core = "^0.3.10"
langchain = "^0.3.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
